+++ src/index.ts
@@ -5,7 +5,7 @@ export const uuidv4 = (): string => {
    );
 };
 
-export const shortIdGenerator = (length: number = 7): string => {
+export const shortId = (length = 7): string => {
    const zerosString = Object.values(new Uint32Array(length + 1)).join('');
    return String(
       Math.floor(Math.random() * Number(`1${zerosString}`)).toString(16)